move app from src to apps and make it use base app class; bump version
