+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "moviesdb"
-version = "1.0.3"
+version = "1.0.4"
 requires-python = ">=3.13"
 dependencies = [
     "boto3>=1.40.15",
