+++ src/apps/__init__.py
@@ -1,5 +1,6 @@
 from .sqlapp import SqlApp
 from .book import BooksApp
 from .image import ImagesApp
+from .app import App
 
-__all__ = ["SqlApp", "BooksApp", "ImagesApp"]
+__all__ = ["SqlApp", "BooksApp", "ImagesApp", "App"]
