+++ src/apps/base.py
@@ -1,5 +1,6 @@
 from abc import ABC, abstractmethod
 import os
+import logging
 
 from collections.abc import Callable
 
@@ -8,6 +9,13 @@ from src.parser import ParsingError, PositionalArgs, KeywordArgs, Flags, parse
 from src.utils.help_utils import parse_docstring, get_rich_help
 from src.utils.utils import possible_match
 
+
+logger = logging.getLogger(__name__)
+
+
+DEFAULT_COMMAND_ALIASES: dict[str, str] = {"clear": "cls"}
+
+# TODO more logging in this module
 
 class BaseApp(ABC):
     def __init__(
@@ -19,9 +27,7 @@ class BaseApp(ABC):
         self.cns = cns
         self.input = input_fn
         self.prompt_str = prompt_str
-
         self.running = True
-
         self.command_methods: dict[
             str, Callable[[PositionalArgs, KeywordArgs, Flags], None]
         ] = {
@@ -29,11 +35,16 @@ class BaseApp(ABC):
             for method_name in dir(self)
             if method_name.startswith("cmd_")
         }
-
         self.help_messages = {
             cmd_root: parse_docstring(cmd_fn.__doc__)
             for cmd_root, cmd_fn in self.command_methods.items()
         }
+        self.register_aliases(DEFAULT_COMMAND_ALIASES)
+
+    def register_aliases(self, aliases: dict[str, str]):
+        for alias, command in aliases.items():
+            self.command_methods[alias] = self.command_methods[command]
+            self.help_messages[alias] = self.help_messages[command]
 
     def cmd_exit(self, pos: PositionalArgs, kwargs: KeywordArgs, flags: Flags):
         """exit
@@ -68,23 +79,39 @@ class BaseApp(ABC):
             + 'Type "help" for a list of commands'
         )
 
+    def process_command(self, command: str):
+        try:
+            root, pos, kwargs, flags = parse(command)
+        except ParsingError as e:
+            self.error(f"{e}: {command!r}")
+            logger.info(f"parsing error: {e} for command {command!r}")
+            return
+        command_method = self.command_methods.get(root)
+        if command_method is None:
+            self._maybe_command(root)
+            return
+        if "help" in flags:
+            self.cmd_help([root], {}, set())
+            return
+        command_method(pos, kwargs, flags)
+        logger.info(f"executed command: {root=!r}, {pos=!r}, {kwargs=!r}, {flags=!r}")
+
     def run(self):
+        logger.info("starting App")
         self.pre_run()
         while self.running:
-            command = self.input(f"[bold yellow]{self.prompt_str} ")
             try:
-                root, pos, kwargs, flags = parse(command)
-            except ParsingError as e:
-                self.cns.print(f"{e}: {command!r}", style="bold red")
-                continue
-            command_method = self.command_methods.get(root)
-            if command_method is None:
-                self._maybe_command(root)
-                continue
-            if "help" in flags:
-                self.cmd_help([root], {}, set())
-                continue
-            command_method(pos, kwargs, flags)
+                command = self.input(self.prompt_str + " ")
+                self.process_command(command)
+            except EOFError:
+                print()
+                return
+            except KeyboardInterrupt:
+                return
+            except Exception as _:
+                self.cns.print_exception()
+                logger.error("unhandled exception", exc_info=True)
+        logger.info("stopping App")
         self.post_run()
 
     @abstractmethod
